+++ src/publishers.cpp
@@ -152,6 +152,18 @@ const char* ToString(Venue venue) {
     case Venue::Equs: {
       return "EQUS";
     }
+    case Venue::Ifus: {
+      return "IFUS";
+    }
+    case Venue::Ifll: {
+      return "IFLL";
+    }
+    case Venue::Xeur: {
+      return "XEUR";
+    }
+    case Venue::Xeer: {
+      return "XEER";
+    }
     default: {
       return "Unknown";
     }
@@ -306,6 +318,18 @@ Venue FromString(const std::string& str) {
   if (str == "EQUS") {
     return Venue::Equs;
   }
+  if (str == "IFUS") {
+    return Venue::Ifus;
+  }
+  if (str == "IFLL") {
+    return Venue::Ifll;
+  }
+  if (str == "XEUR") {
+    return Venue::Xeur;
+  }
+  if (str == "XEER") {
+    return Venue::Xeer;
+  }
   throw InvalidArgumentError{"FromString<Venue>", "str",
                              "unknown value '" + str + '\''};
 }
@@ -416,6 +440,18 @@ const char* ToString(Dataset dataset) {
     }
     case Dataset::EqusMini: {
       return "EQUS.MINI";
+    }
+    case Dataset::IfusImpact: {
+      return "IFUS.IMPACT";
+    }
+    case Dataset::IfllImpact: {
+      return "IFLL.IMPACT";
+    }
+    case Dataset::XeurEobi: {
+      return "XEUR.EOBI";
+    }
+    case Dataset::XeerEobi: {
+      return "XEER.EOBI";
     }
     default: {
       return "Unknown";
@@ -534,6 +570,18 @@ Dataset FromString(const std::string& str) {
   }
   if (str == "EQUS.MINI") {
     return Dataset::EqusMini;
+  }
+  if (str == "IFUS.IMPACT") {
+    return Dataset::IfusImpact;
+  }
+  if (str == "IFLL.IMPACT") {
+    return Dataset::IfllImpact;
+  }
+  if (str == "XEUR.EOBI") {
+    return Dataset::XeurEobi;
+  }
+  if (str == "XEER.EOBI") {
+    return Dataset::XeerEobi;
   }
   throw InvalidArgumentError{"FromString<Dataset>", "str",
                              "unknown value '" + str + '\''};
@@ -828,6 +876,24 @@ Venue PublisherVenue(Publisher publisher) {
     }
     case Publisher::XnysTradesEqus: {
       return Venue::Equs;
+    }
+    case Publisher::IfusImpactIfus: {
+      return Venue::Ifus;
+    }
+    case Publisher::IfusImpactXoff: {
+      return Venue::Xoff;
+    }
+    case Publisher::IfllImpactIfll: {
+      return Venue::Ifll;
+    }
+    case Publisher::IfllImpactXoff: {
+      return Venue::Xoff;
+    }
+    case Publisher::XeurEobiXeur: {
+      return Venue::Xeur;
+    }
+    case Publisher::XeerEobiXeer: {
+      return Venue::Xeer;
     }
     default: {
       throw InvalidArgumentError{
@@ -1127,6 +1193,24 @@ Dataset PublisherDataset(Publisher publisher) {
     case Publisher::XnysTradesEqus: {
       return Dataset::XnysTrades;
     }
+    case Publisher::IfusImpactIfus: {
+      return Dataset::IfusImpact;
+    }
+    case Publisher::IfusImpactXoff: {
+      return Dataset::IfusImpact;
+    }
+    case Publisher::IfllImpactIfll: {
+      return Dataset::IfllImpact;
+    }
+    case Publisher::IfllImpactXoff: {
+      return Dataset::IfllImpact;
+    }
+    case Publisher::XeurEobiXeur: {
+      return Dataset::XeurEobi;
+    }
+    case Publisher::XeerEobiXeer: {
+      return Dataset::XeerEobi;
+    }
     default: {
       throw InvalidArgumentError{
           "PublisherDataset", "publisher",
@@ -1425,6 +1509,24 @@ const char* ToString(Publisher publisher) {
     }
     case Publisher::XnysTradesEqus: {
       return "XNYS.TRADES.EQUS";
+    }
+    case Publisher::IfusImpactIfus: {
+      return "IFUS.IMPACT.IFUS";
+    }
+    case Publisher::IfusImpactXoff: {
+      return "IFUS.IMPACT.XOFF";
+    }
+    case Publisher::IfllImpactIfll: {
+      return "IFLL.IMPACT.IFLL";
+    }
+    case Publisher::IfllImpactXoff: {
+      return "IFLL.IMPACT.XOFF";
+    }
+    case Publisher::XeurEobiXeur: {
+      return "XEUR.EOBI.XEUR";
+    }
+    case Publisher::XeerEobiXeer: {
+      return "XEER.EOBI.XEER";
     }
     default: {
       return "Unknown";
@@ -1726,6 +1828,24 @@ Publisher FromString(const std::string& str) {
   }
   if (str == "XNYS.TRADES.EQUS") {
     return Publisher::XnysTradesEqus;
+  }
+  if (str == "IFUS.IMPACT.IFUS") {
+    return Publisher::IfusImpactIfus;
+  }
+  if (str == "IFUS.IMPACT.XOFF") {
+    return Publisher::IfusImpactXoff;
+  }
+  if (str == "IFLL.IMPACT.IFLL") {
+    return Publisher::IfllImpactIfll;
+  }
+  if (str == "IFLL.IMPACT.XOFF") {
+    return Publisher::IfllImpactXoff;
+  }
+  if (str == "XEUR.EOBI.XEUR") {
+    return Publisher::XeurEobiXeur;
+  }
+  if (str == "XEER.EOBI.XEER") {
+    return Publisher::XeerEobiXeer;
   }
   throw InvalidArgumentError{"FromString<Publisher>", "str",
                              "unknown value '" + str + '\''};